Fix last commit (IN PROGRESS)
Last commit didn't really function correctly, due to a misinterpretation of the
results. This fixes that as well as handling the cases when the number of
workers is more than the number of columns.

However, a concurrency problem is still present and occurs only some of the
times. The higher the number of workers is, the more likely this will happen.

This might be fixed in some later revisions.