Fix most bugs
The program works in most cases now. Only minor exception is when the number of
workers is greater than the total number of entries.

Memory allocated is now freed after use. Program works 100% of the times when
it's not the exception case, even when the number of workers is greater than the
dimension of the matrices (but still less than the total number of entries).